+++ lib/src/widget/provider/provider_widget.dart
@@ -14,38 +14,53 @@ class BaseProviderWidget<T extends ChangeNotifier, Theme, Localization>
     Theme theme,
     Localization localization,
   )? builderWithThemeAndLocalizations;
-  final T Function() create;
+  final T Function()? create;
+  final T? value;
 
   const BaseProviderWidget({
-    required this.create,
+    required T Function() this.create,
     this.builder,
     this.builderWithThemeAndLocalizations,
     super.key,
-  });
+  }) : value = null;
+
+  const BaseProviderWidget.value({
+    required this.value,
+    this.builder,
+    this.builderWithThemeAndLocalizations,
+    super.key,
+  }) : create = null;
 
   @override
   Widget build(BuildContext context) {
+    final child = Consumer<T>(
+      builder: (context, viewModel, child) {
+        if (builder != null) {
+          return builder!.call(
+            context,
+            viewModel,
+          );
+        } else if (builderWithThemeAndLocalizations != null) {
+          return builderWithThemeAndLocalizations!(
+            context,
+            viewModel,
+            themeLookup(context),
+            localizationLookup(context),
+          );
+        }
+        throw ArgumentError(
+            'builder and builderWithThemeAndLocalizations are null');
+      },
+    );
+    if (value != null) {
+      return ChangeNotifierProvider<T>.value(
+        value: value!,
+        child: child,
+      );
+    }
     return ChangeNotifierProvider<T>(
-      create: (context) => create(),
-      child: Consumer<T>(
-        builder: (context, viewModel, child) {
-          if (builder != null) {
-            return builder!.call(
-              context,
-              viewModel,
-            );
-          } else if (builderWithThemeAndLocalizations != null) {
-            return builderWithThemeAndLocalizations!(
-              context,
-              viewModel,
-              themeLookup(context),
-              localizationLookup(context),
-            );
-          }
-          throw ArgumentError(
-              'builder and builderWithThemeAndLocalizations are null');
-        },
-      ),
+      create: (context) => create!(),
+      child: child,
     );
   }
 }